move sentryapp model to sentry_apps dir
