+++ src/sentry/api/endpoints/integrations/sentry_apps/internal_app_token/index.py
@@ -13,9 +13,9 @@ from sentry.api.endpoints.integrations.sentry_apps.details import (
 from sentry.api.serializers.models.apitoken import ApiTokenSerializer
 from sentry.exceptions import ApiTokenLimitError
 from sentry.models.apitoken import ApiToken
-from sentry.models.integrations.sentry_app import MASKED_VALUE
 from sentry.models.integrations.sentry_app_installation import SentryAppInstallation
 from sentry.sentry_apps.installations import SentryAppInstallationTokenCreator
+from sentry.sentry_apps.models.sentry_app import MASKED_VALUE
 
 
 @control_silo_endpoint
