+++ src/sentry/api/serializers/models/sentry_app.py
@@ -13,8 +13,8 @@ from sentry.hybridcloud.services.organization_mapping import organization_mappin
 from sentry.integrations.models.integration_feature import IntegrationFeature, IntegrationTypes
 from sentry.models.apiapplication import ApiApplication
 from sentry.models.avatars.sentry_app_avatar import SentryAppAvatar
-from sentry.models.integrations.sentry_app import MASKED_VALUE, SentryApp
 from sentry.organizations.services.organization import organization_service
+from sentry.sentry_apps.models.sentry_app import MASKED_VALUE, SentryApp
 from sentry.users.models.user import User
 from sentry.users.services.user.service import user_service
