+++ src/sentry/api/serializers/rest_framework/sentry_app.py
@@ -7,7 +7,7 @@ from sentry.api.serializers.rest_framework.base import camel_to_snake_case
 from sentry.api.validators.sentry_apps.schema import validate_ui_element_schema
 from sentry.integrations.models.integration_feature import Feature
 from sentry.models.apiscopes import ApiScopes
-from sentry.models.integrations.sentry_app import (
+from sentry.sentry_apps.models.sentry_app import (
     REQUIRED_EVENT_PERMISSIONS,
     UUID_CHARS_IN_SLUG,
     VALID_EVENT_RESOURCES,
