+++ src/sentry/api/serializers/rest_framework/sentry_app_request.py
@@ -7,9 +7,9 @@ from django.urls import reverse
 
 from sentry import eventstore
 from sentry.api.serializers import Serializer
-from sentry.models.integrations.sentry_app import SentryApp
 from sentry.models.organization import Organization
 from sentry.models.project import Project
+from sentry.sentry_apps.models.sentry_app import SentryApp
 from sentry.utils.sentry_apps.webhooks import TIMEOUT_STATUS_CODE
