+++ src/sentry/conf/server.py
@@ -397,6 +397,7 @@ INSTALLED_APPS: tuple[str, ...] = (
     "sentry.nodestore",
     "sentry.users",
     "sentry.integrations",
+    "sentry.sentry_apps",
     "sentry.monitors",
     "sentry.uptime",
     "sentry.replays",
