+++ src/sentry/mediators/external_requests/util.py
@@ -6,7 +6,7 @@ from requests.exceptions import ConnectionError, Timeout
 from requests.models import Response
 
 from sentry.http import safe_urlopen
-from sentry.models.integrations.sentry_app import SentryApp, track_response_code
+from sentry.sentry_apps.models.sentry_app import SentryApp, track_response_code
 from sentry.utils.sentry_apps import SentryAppWebhookRequestsBuffer
 from sentry.utils.sentry_apps.webhooks import TIMEOUT_STATUS_CODE
