+++ src/sentry/mediators/sentry_app_installations/installation_notifier.py
@@ -6,8 +6,8 @@ from sentry.coreapi import APIUnauthorized
 from sentry.mediators.mediator import Mediator
 from sentry.mediators.param import Param
 from sentry.models.apigrant import ApiGrant
-from sentry.models.integrations.sentry_app import SentryApp
 from sentry.models.integrations.sentry_app_installation import SentryAppInstallation
+from sentry.sentry_apps.models.sentry_app import SentryApp
 from sentry.users.services.user.model import RpcUser
 from sentry.utils.sentry_apps import send_and_save_webhook_request
