+++ src/sentry/models/integrations/__init__.py
@@ -9,10 +9,10 @@ __all__ = (
 
 # REQUIRED for migrations to run.
 from sentry.integrations.types import ExternalProviders  # NOQA
-from sentry.models.integrations.sentry_app import SentryApp
 from sentry.models.integrations.sentry_app_component import SentryAppComponent
 from sentry.models.integrations.sentry_app_installation import SentryAppInstallation
 from sentry.models.integrations.sentry_app_installation_for_provider import (
     SentryAppInstallationForProvider,
 )
 from sentry.models.integrations.sentry_app_installation_token import SentryAppInstallationToken
+from sentry.sentry_apps.models.sentry_app import SentryApp
