+++ src/sentry/receivers/outbox/control.py
@@ -22,12 +22,12 @@ from sentry.integrations.models.integration import Integration
 from sentry.issues.services.issue import issue_service
 from sentry.models.apiapplication import ApiApplication
 from sentry.models.files.utils import get_relocation_storage
-from sentry.models.integrations.sentry_app import SentryApp
 from sentry.models.integrations.sentry_app_installation import SentryAppInstallation
 from sentry.models.organizationmapping import OrganizationMapping
 from sentry.organizations.services.organization import RpcOrganizationSignal, organization_service
 from sentry.receivers.outbox import maybe_process_tombstone
 from sentry.relocation.services.relocation_export.service import region_relocation_export_service
+from sentry.sentry_apps.models.sentry_app import SentryApp
 from sentry.sentry_apps.services.app.service import get_by_application_id, get_installation
 
 logger = logging.getLogger(__name__)
