+++ src/sentry/sentry_apps/services/app/serial.py
@@ -2,8 +2,8 @@ from sentry.constants import SentryAppStatus
 from sentry.models.apiapplication import ApiApplication
 from sentry.models.apitoken import ApiToken
 from sentry.models.integrations import SentryAppComponent
-from sentry.models.integrations.sentry_app import SentryApp
 from sentry.models.integrations.sentry_app_installation import SentryAppInstallation
+from sentry.sentry_apps.models.sentry_app import SentryApp
 from sentry.sentry_apps.services.app import (
     RpcApiApplication,
     RpcSentryApp,
