+++ tests/sentry/deletions/test_sentry_app.py
@@ -3,8 +3,8 @@ from django.db import connections, router
 
 from sentry import deletions
 from sentry.models.apiapplication import ApiApplication
-from sentry.models.integrations.sentry_app import SentryApp
 from sentry.models.integrations.sentry_app_installation import SentryAppInstallation
+from sentry.sentry_apps.models.sentry_app import SentryApp
 from sentry.testutils.cases import TestCase
 from sentry.testutils.silo import control_silo_test
 from sentry.users.models.user import User
