+++ tests/sentry/mediators/token_exchange/test_validator.py
@@ -4,7 +4,7 @@ import pytest
 
 from sentry.coreapi import APIUnauthorized
 from sentry.mediators.token_exchange.validator import Validator
-from sentry.models.integrations.sentry_app import SentryApp
+from sentry.sentry_apps.models.sentry_app import SentryApp
 from sentry.sentry_apps.services.app import app_service
 from sentry.testutils.cases import TestCase
 from sentry.testutils.silo import control_silo_test
