+++ tests/sentry/sentry_apps/test_sentry_app_updater.py
@@ -6,10 +6,10 @@ from django.utils import timezone
 from sentry.constants import SentryAppStatus
 from sentry.coreapi import APIError
 from sentry.models.apitoken import ApiToken
-from sentry.models.integrations.sentry_app import SentryApp
 from sentry.models.integrations.sentry_app_component import SentryAppComponent
 from sentry.models.servicehook import ServiceHook
 from sentry.sentry_apps.apps import SentryAppUpdater, expand_events
+from sentry.sentry_apps.models.sentry_app import SentryApp
 from sentry.silo.base import SiloMode
 from sentry.testutils.cases import TestCase
 from sentry.testutils.silo import assume_test_silo_mode, control_silo_test
