+++ tests/sentry/tasks/test_sentry_apps.py
@@ -17,9 +17,9 @@ from sentry.integrations.notify_disable import notify_disable
 from sentry.integrations.request_buffer import IntegrationRequestBuffer
 from sentry.models.activity import Activity
 from sentry.models.auditlogentry import AuditLogEntry
-from sentry.models.integrations.sentry_app import SentryApp
 from sentry.models.integrations.sentry_app_installation import SentryAppInstallation
 from sentry.models.rule import Rule
+from sentry.sentry_apps.models.sentry_app import SentryApp
 from sentry.shared_integrations.exceptions import ClientError
 from sentry.tasks.post_process import post_process_group
 from sentry.tasks.sentry_apps import (
